form_builder/src/requirements.txt
transformers==4.28.1
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
python-dotenv>=1.0.0
torch>=2.1.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
protobuf>=4.25.1
sentencepiece>=0.1.99 